Init load hooks. Add Shell.Exec.
Adds postload and preunload hooks that run in the current interactive
shell. This adds two environment variables DIRENV_POSTLOAD and
DIRENV_PREUNLOAD which are managed by an envrc. The changing of these
environment variables are directly outputted by `direnv export` at the
necessary times so that the running shell may `eval` them.